adsa

+++ circle/sample/WORKING_BUILD/code_parser.cpp
@@ -170,7 +170,7 @@ bool            CKernel::parse264                  (   h264_state* h,
                                 filename_array[file_index - p_fromFile], EMPTYLOG,
                                 "FileSize", (u32)size_array[file_index - p_fromFile]);
 
-                    storeLog( MY_BUFFER, MY_INDEX "blockbase", blockBase, "buffer address", data, "index", file_index );
+                    storeLog( MY_BUFFER, MY_INDEX "blockbase", blockBase, "buffer address", (u32)data, "index", file_index );
 
                     storeMsg(   MY_BUFFER, MY_INDEX, "vid_file first 1024 byte", p_buffer_array, 1024 );
